lambda/requirements.txt
boto3==1.20.22
botocore==1.23.22
certifi==2021.10.8
charset-normalizer==2.0.9
idna==3.3
jmespath==0.10.0
pycryptodomex==3.12.0
PyMySQL==1.0.2
python-dateutil==2.8.2
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.7
